+++ LexiconVendingMachine/ProductToy.cs
@@ -4,7 +4,7 @@ namespace LexiconVendingMachine
 {
     class ProductToy : Product
     {
-        public ProductToy(string name, decimal price, int stock)
+        public ProductToy(string name, int price, int stock)
         {
             this.Name = name;
             this.Size = 1;
@@ -12,9 +12,7 @@ namespace LexiconVendingMachine
             this.Unit = "pcs";
             this.InStock = stock;
         }
-
         
-
         public override string Use()
         {
             string instructions = "Play and have fun...";
